requirements.txt
APScheduler==3.10.4
beautifulsoup4==4.12.2
blinker==1.6.2
bs4==0.0.1
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
ElasticEmail==4.0.23
Flask==2.3.3
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
mysql-connector-python==8.1.0
protobuf==4.21.12
python-dateutil==2.8.2
pytz==2023.3
requests==2.31.0
six==1.16.0
soupsieve==2.4.1
tzlocal==5.0.1
urllib3==2.0.4
Werkzeug==2.3.7
